# src/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_boq_api_loans__loan_id__boq_post": {
        "properties": {
          "built_up_sqft": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Built Up Sqft"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_boq_api_loans__loan_id__boq_post",
        "type": "object"
      },
      "BoqReviewView": {
        "properties": {
          "all_groups": {
            "items": {
              "$ref": "#/components/schemas/FlagGroupView"
            },
            "title": "All Groups",
            "type": "array"
          },
          "amount_before_slab": {
            "title": "Amount Before Slab",
            "type": "number"
          },
          "analysis_mode": {
            "default": "unknown",
            "title": "Analysis Mode",
            "type": "string"
          },
          "boq_total": {
            "title": "Boq Total",
            "type": "number"
          },
          "borrower": {
            "title": "Borrower",
            "type": "string"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/StatCardView"
            },
            "title": "Cards",
            "type": "array"
          },
          "contractor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contractor"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/FlagGroupView"
            },
            "title": "Groups",
            "type": "array"
          },
          "gst_stated": {
            "title": "Gst Stated",
            "type": "boolean"
          },
          "item_count": {
            "title": "Item Count",
            "type": "integer"
          },
          "loan_id": {
            "title": "Loan Id",
            "type": "string"
          },
          "payment_schedule": {
            "items": {
              "$ref": "#/components/schemas/PaymentStageView"
            },
            "title": "Payment Schedule",
            "type": "array"
          },
          "pct_before_slab": {
            "title": "Pct Before Slab",
            "type": "number"
          },
          "provenance": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Provenance",
            "type": "object"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionView"
            },
            "title": "Questions",
            "type": "array"
          },
          "received_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received On"
          },
          "rev": {
            "title": "Rev",
            "type": "integer"
          }
        },
        "required": [
          "loan_id",
          "analysis_mode",
          "provenance",
          "borrower",
          "contractor",
          "received_on",
          "item_count",
          "rev",
          "cards",
          "groups",
          "all_groups",
          "questions",
          "payment_schedule",
          "boq_total",
          "pct_before_slab",
          "amount_before_slab",
          "gst_stated"
        ],
        "title": "BoqReviewView",
        "type": "object"
      },
...
      "FlagGroupView": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FlagRowView"
            },
            "title": "Items",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "items"
        ],
        "title": "FlagGroupView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobStatusResponse": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "loan_id": {
            "title": "Loan Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "loan_id",
          "status",
          "error"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "loan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loan Id"
          },
          "phone": {
            "pattern": "^\\d{10}$",
            "title": "Phone",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "bank"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "phone"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "loan_id": {
            "title": "Loan Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "bank"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "loan_id",
          "name"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "PaymentStageView": {
        "properties": {
          "before_slab": {
            "title": "Before Slab",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "pct": {
            "title": "Pct",
            "type": "number"
          }
        },
        "required": [
          "label",
          "pct",
          "before_slab"
        ],
        "title": "PaymentStageView",
        "type": "object"
      },
...
      "QuestionView": {
        "properties": {
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "number",
          "text",
          "status"
        ],
        "title": "QuestionView",
        "type": "object"
      },
...
      "SanctionBarView": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "pct_of_max": {
            "title": "Pct Of Max",
            "type": "number"
          },
          "sub": {
            "title": "Sub",
            "type": "string"
          },
          "tone": {
            "default": "neutral",
            "enum": [
              "danger",
              "warn",
              "success",
              "neutral"
            ],
            "title": "Tone",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "label",
          "value",
          "pct_of_max",
          "sub",
          "tone"
        ],
        "title": "SanctionBarView",
        "type": "object"
      },
      "SanctionCheckView": {
        "properties": {
          "bars": {
            "items": {
              "$ref": "#/components/schemas/SanctionBarView"
            },
            "title": "Bars",
            "type": "array"
          },
          "loan_id": {
            "title": "Loan Id",
            "type": "string"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/SanctionOptionView"
            },
            "title": "Options",
            "type": "array"
          },
          "provenance": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Provenance",
            "type": "object"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SanctionSectionView"
            },
            "title": "Sections",
            "type": "array"
          },
          "shortfall": {
            "title": "Shortfall",
            "type": "number"
          }
        },
        "required": [
          "loan_id",
          "provenance",
          "bars",
          "shortfall",
          "sections",
          "options"
        ],
        "title": "SanctionCheckView",
        "type": "object"
      },
      "SanctionOptionView": {
        "properties": {
          "desc": {
            "title": "Desc",
            "type": "string"
          },
          "saves_label": {
            "title": "Saves Label",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "saves_label",
          "desc"
        ],
        "title": "SanctionOptionView",
        "type": "object"
      },
      "SanctionSectionView": {
        "properties": {
          "delta": {
            "title": "Delta",
            "type": "number"
          },
          "market": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "quoted": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quoted"
          },
          "quoted_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quoted Note"
          },
          "tone": {
            "enum": [
              "danger",
              "warn",
              "success",
              "neutral"
            ],
            "title": "Tone",
            "type": "string"
          }
        },
        "required": [
          "name",
          "quoted",
          "quoted_note",
          "market",
          "delta",
          "tone"
        ],
        "title": "SanctionSectionView",
        "type": "object"
      },
...
      "StatCardView": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "sub": {
            "title": "Sub",
            "type": "string"
          },
          "tone": {
            "default": "neutral",
            "enum": [
              "danger",
              "warn",
              "success",
              "neutral"
            ],
            "title": "Tone",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          },
          "value_kind": {
            "enum": [
              "money",
              "money_compact",
              "count",
              "pct",
              "ratio",
              "text"
            ],
            "title": "Value Kind",
            "type": "string"
          }
        },
        "required": [
          "label",
          "value",
          "value_kind",
          "sub",
          "tone"
        ],
        "title": "StatCardView",
        "type": "object"
      },
...
      "UploadAccepted": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "loan_id": {
            "title": "Loan Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "loan_id"
        ],
        "title": "UploadAccepted",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Serves pipeline-shaped loan data to the Neev owner and bank consoles.",
    "title": "Neev API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/session": {
      "delete": {
        "operationId": "delete_session_api_auth_session_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Delete Session",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "operationId": "create_session_api_auth_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/jobs/{job_id}": {
      "get": {
        "description": "Why a run failed, for a client that missed the stream's error event.",
        "operationId": "job_status_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Status",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/events": {
      "get": {
        "operationId": "job_events_api_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Events",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/loans/{loan_id}/boq": {
      "post": {
        "operationId": "upload_boq_api_loans__loan_id__boq_post",
        "parameters": [
          {
            "description": "The loan's id, e.g. 1001.",
            "in": "path",
            "name": "loan_id",
            "required": true,
            "schema": {
              "description": "The loan's id, e.g. 1001.",
              "title": "Loan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_boq_api_loans__loan_id__boq_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Boq",
        "tags": [
          "boq"
        ]
      }
    },
    "/api/loans/{loan_id}/boq/latest": {
      "get": {
        "operationId": "boq_latest_api_loans__loan_id__boq_latest_get",
        "parameters": [
          {
            "description": "The loan's id, e.g. 1001.",
            "in": "path",
            "name": "loan_id",
            "required": true,
            "schema": {
              "description": "The loan's id, e.g. 1001.",
              "title": "Loan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoqReviewView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Boq Latest",
        "tags": [
          "boq"
        ]
      }
    },
    "/api/loans/{loan_id}/boq/rev/{rev}": {
      "get": {
        "operationId": "boq_revision_api_loans__loan_id__boq_rev__rev__get",
        "parameters": [
          {
            "in": "path",
            "name": "rev",
            "required": true,
            "schema": {
              "title": "Rev",
              "type": "integer"
            }
          },
          {
            "description": "The loan's id, e.g. 1001.",
            "in": "path",
            "name": "loan_id",
            "required": true,
            "schema": {
              "description": "The loan's id, e.g. 1001.",
              "title": "Loan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoqReviewView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Boq Revision",
        "tags": [
          "boq"
        ]
      }
    },
...
    "/api/loans/{loan_id}/sanction-check": {
      "get": {
        "operationId": "sanction_check_api_loans__loan_id__sanction_check_get",
        "parameters": [
          {
            "description": "The loan's id, e.g. 1001.",
            "in": "path",
            "name": "loan_id",
            "required": true,
            "schema": {
              "description": "The loan's id, e.g. 1001.",
              "title": "Loan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanctionCheckView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sanction Check",
        "tags": [
          "boq"
        ]
      }
    },
...
  }
}